% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealReadSim.R
\name{addToDataSystem}
\alias{addToDataSystem}
\title{addToDataSystem}
\usage{
addToDataSystem(
  filenames_csv,
  bowtieOptions = "--no-unal",
  minIdL,
  metagenomeDir,
  readAsBams,
  bowtiebuildOptions,
  threads = 1
)
}
\arguments{
\item{filenames_csv}{A character vector specifying the files containing the paths pointing to the fasta files and bam/fastq files of the input.}

\item{bowtieOptions}{A character vector containing the bowtie2 options for building a datasystem.}

\item{minIdL}{The minimum length of a sequence that is identical on two genomes to be considered.}

\item{metagenomeDir}{A String specifying which metagenome datasystem to use, or where to build one if filenames_csv is given.}

\item{readAsBams}{A boolean value determening wether filenames_csv should be read as containing bam file names or fastq.}

\item{bowtiebuildOptions}{A character vector containing the bowtie2-build options for building a datasystem.}
}
\value{
This function has no return value
}
\description{
This function sets up the datasystems for realReadSim
}
