% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RealReadSim.R
\name{realReadSim}
\alias{realReadSim}
\title{realReadSim}
\usage{
realReadSim(
  filenames_csv,
  coverage,
  bowtieOptions = "--no-unal",
  humanReadable = FALSE,
  readAsBams = TRUE,
  minMapq = 40,
  redraw = FALSE,
  repeatable = TRUE,
  seed = 0,
  minContigLength = 10000
)
}
\arguments{
\item{filenames_csv}{This parameter takes the name of a file in which the names of the fasta files and the name of the bam or fastq files and or the bowtie prebuild name are contained}

\item{coverage}{A integer vector containing the wished coverages of the individual genomes}

\item{humanReadable}{A boolean value determening wether there will be graphical output or not}

\item{readAsBams}{A boolean value determening wether filenames_csv should be read as containing bam file names or not}

\item{minMapq}{A integer value stating the minimal mapq score of reads to use}

\item{redraw}{A boolean value determening, if the asked for coverage is to much for the data set, reads from the same data set should be drawn again}

\item{repeatable}{A boolean value determening wether a seed will be set}

\item{seed}{A integer value used as seed when reapeatable = TRUE}

\item{prebuild}{A boolean value that determines wether the prebuild function is to be used}

\item{bowtieDefaultOptions}{A boolean value determening wether the set bowtie2 options schould be used or user defined options}

\item{asTempoaryDatabase}{A boolean value determening wether the data will be read in all at once or one by one}
}
\description{
This function simulates metagenomic data using real experimental reads
}
